# packages/api-client/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExecutionEvent": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventError"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "result_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduler_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchedulerRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "enum": [
              "accepted",
              "running",
              "completed",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "status_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Detail"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "tenant_id",
          "workspace_id",
          "job_id",
          "submission_id",
          "execution_id",
          "state",
          "occurred_at",
          "trace_id"
        ],
        "title": "ExecutionEvent",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuntimeEventAck": {
        "additionalProperties": false,
        "properties": {
          "idempotent": {
            "default": false,
            "title": "Idempotent",
            "type": "boolean"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "RuntimeEventAck",
        "type": "object"
      },
      "RuntimeJobStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "execution_owner": {
            "title": "Execution Owner",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "accepted",
              "running",
              "completed",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "submission_id",
          "execution_owner",
          "state",
          "updated_at",
          "trace_id"
        ],
        "title": "RuntimeJobStatusResponse",
        "type": "object"
      },
      "RuntimeNodeJoinTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "node_name_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Name Hint"
          },
          "queue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Name"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds"
          }
        },
        "title": "RuntimeNodeJoinTokenRequest",
        "type": "object"
      },
      "RuntimeNodeJoinTokenResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "install_command": {
            "title": "Install Command",
            "type": "string"
          },
          "install_script_url": {
            "title": "Install Script Url",
            "type": "string"
          },
          "join_token": {
            "title": "Join Token",
            "type": "string"
          },
          "queue_name": {
            "title": "Queue Name",
            "type": "string"
          },
          "register_endpoint": {
            "title": "Register Endpoint",
            "type": "string"
          },
          "token_ttl_seconds": {
            "title": "Token Ttl Seconds",
            "type": "integer"
          }
        },
        "required": [
          "join_token",
          "expires_at",
          "token_ttl_seconds",
          "queue_name",
          "install_script_url",
          "register_endpoint",
          "install_command"
        ],
        "title": "RuntimeNodeJoinTokenResponse",
        "type": "object"
      },
      "RuntimeNodeRegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "RuntimeNodeRegisterRequest",
        "type": "object"
      },
      "RuntimeNodeRegisterResponse": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "queue_name": {
            "title": "Queue Name",
            "type": "string"
          },
          "registered_at": {
            "title": "Registered At",
            "type": "string"
          },
          "server_id": {
            "title": "Server Id",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "server_id",
          "target_id",
          "queue_name",
          "registered_at"
        ],
        "title": "RuntimeNodeRegisterResponse",
        "type": "object"
      },
...
      "SubmitJobAccepted": {
        "additionalProperties": false,
        "properties": {
          "accepted_at": {
            "title": "Accepted At",
            "type": "string"
          },
          "execution_owner": {
            "title": "Execution Owner",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "submission_id",
          "execution_owner",
          "accepted_at",
          "trace_id"
        ],
        "title": "SubmitJobAccepted",
        "type": "object"
      },
      "SubmitJobCommand": {
        "additionalProperties": false,
        "properties": {
          "execution_profile": {
            "$ref": "#/components/schemas/SubmitExecutionProfile"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "management_node_id": {
            "title": "Management Node Id",
            "type": "string"
          },
          "payload_ref": {
            "$ref": "#/components/schemas/SubmitPayloadRef"
          },
          "requested_by": {
            "$ref": "#/components/schemas/SubmitRequestedBy"
          },
          "resource_shape": {
            "$ref": "#/components/schemas/SubmitResourceShape"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "workspace_id",
          "job_id",
          "idempotency_key",
          "management_node_id",
          "execution_profile",
          "resource_shape",
          "payload_ref",
          "requested_by"
        ],
        "title": "SubmitJobCommand",
        "type": "object"
      },
...
      "ZPEJobStatus": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "enum": [
              "queued",
              "started",
              "finished",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "status"
        ],
        "title": "ZPEJobStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chem Model API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runtime/jobs/{job_id}": {
      "get": {
        "operationId": "get_runtime_job_api_runtime_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeJobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runtime Job",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/jobs/{job_id}/detail": {
      "get": {
        "operationId": "get_runtime_job_detail_api_runtime_jobs__job_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Runtime Job Detail Api Runtime Jobs  Job Id  Detail Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runtime Job Detail",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/jobs/{job_id}/events": {
      "post": {
        "operationId": "post_runtime_event_api_runtime_jobs__job_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeEventAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Runtime Event",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/jobs/{job_id}/projection": {
      "get": {
        "operationId": "get_runtime_job_projection_api_runtime_jobs__job_id__projection_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZPEJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runtime Job Projection",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/jobs:submit": {
      "post": {
        "operationId": "submit_runtime_job_api_runtime_jobs_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitJobCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitJobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Runtime Job",
        "tags": [
          "runtime"
        ]
      }
    },
...
    "/api/runtime/nodes/join-token": {
      "post": {
        "operationId": "issue_runtime_node_join_token_api_runtime_nodes_join_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeNodeJoinTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeNodeJoinTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Runtime Node Join Token",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/nodes/register": {
      "post": {
        "operationId": "register_runtime_compute_node_api_runtime_nodes_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeNodeRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeNodeRegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Runtime Compute Node",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/parse": {
      "post": {
        "operationId": "parse_runtime_input_api_runtime_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZPEParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZPEParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Runtime Input",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/targets": {
      "get": {
        "operationId": "list_runtime_targets_api_runtime_targets_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTargetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runtime Targets",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/runtime/targets/{target_id}/active": {
      "put": {
        "operationId": "select_runtime_target_api_runtime_targets__target_id__active_put",
        "parameters": [
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTargetSelectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Runtime Target",
        "tags": [
          "runtime"
        ]
      }
    },
...
  }
}